+++ src/main/java/cart/dto/purchaseorder/response/PurchaseOrderPageResponse.java
@@ -1,7 +1,5 @@
 package cart.dto.purchaseorder.response;
 
-import cart.dto.purchaseorder.PurchaseOrderItemInfoResponse;
-
 import java.util.List;
 import java.util.Objects;
