+++ src/test/java/cart/integration/PurchaseOrderIntegrationTest.java
@@ -4,9 +4,9 @@ import cart.dao.MemberDao;
 import cart.domain.Member;
 import cart.dto.product.ProductRequest;
 import cart.dto.product.ProductResponse;
-import cart.dto.purchaseorder.PurchaseOrderItemInfoResponse;
 import cart.dto.purchaseorder.request.PurchaseOrderItemRequest;
 import cart.dto.purchaseorder.request.PurchaseOrderRequest;
+import cart.dto.purchaseorder.response.PurchaseOrderItemInfoResponse;
 import cart.dto.purchaseorder.response.PurchaseOrderItemResponse;
 import cart.dto.purchaseorder.response.PurchaseOrderPageResponse;
 import cart.dto.purchaseorder.response.PurchaseOrderResponse;
